+++ src/jinnjs.js
@@ -1,14 +1,11 @@
+var Tasks = require('./tasks');
+
 module.exports = JinnJS;
 
+/**
+ * JinnJS.
+ * @constructor
+ */
 function JinnJS() {
-    this._tasks = {};
+    this.tasks = new Tasks();
 }
-
-JinnJS.prototype.task = function(name, task) {
-    if (this._tasks[name]) { throw new Error('Task with name "' + name + '" already exists.'); }
-    this._tasks[name] = task;
-};
-
-JinnJS.prototype.tasks = function(obj) {
-    Object.keys(obj = obj || {}).forEach(function(key) { this.task(key, obj[key]); }.bind(this));
-};